feat(server): Added server app with typescript and express.js.
This new app is a *server app* made with:

- Node.JS
- Express.JS
- TypeScript

It is a type safe typescript server app. It also includes a config file
where you can put in a type safe way all of your env variables and use
them across your application.

It uses nodemon to start the development server and work locally while
you make changes.

You can use TypeScript "tsconfig.json" paths inside, currently there is
only "@/*".

The "app.ts" and "server.ts" are separated into single files, the most
important file will be the "server.ts". This file will be the one that
is executed in development or production.

Dot files added including:

- .env.example
- .gitignore

Added `prettier` and `eslint` based on the `config/` folder application
that has the general configuration for this utilities.

You can read the `README.md` file for a better explanation.