updated ui shared package for components
With the no longer support for custom components, now there is only a
simple Button component that can be exported.

Removed shadcn/ui related packages.

Replaced prettier and eslint scripts with biome scripts.

Modified the tsconfig.json to include the main base.json export and add
the rest of necessary configuration to it.

Updated the README.md to highlight the changes of shadcn/ui.